topic tracker renaming and changing flaskMain route
