+++ flask_app/HTML/HaroTable/haroTableView.html
@@ -73,7 +73,7 @@
                 </a>
             </div>
             <div>
-                <a href="/jtracker">
+                <a href="/topic_tracker">
                     Journalist Tracker
                 </a>
             </div>
