+++ flask_app/HTML/contactUs.html
@@ -18,7 +18,7 @@
             <ul>
                 <li><a href="/home"> Home </a></li>
                 <li><a href="/email_validator"> Email Verification</a></li>
-                <li><a href="/jtracker"> Periodic Writers </a></li>
+                <li><a href="/topic_tracker"> Periodic Writers </a></li>
                 <li><a href="/haro_table"> Haros Table </a></li>
                 <li><a href="/contact_us"> Contact Us </a></li>
             </ul>
