+++ flask_app/HTML/homePage.html
@@ -25,7 +25,7 @@
                 </a>
             </div>
             <div>
-                <a href="/jtracker">
+                <a href="/topic_tracker">
                     Journalist Tracker
                 </a>
             </div>
