+++ flask_app/HTML/periodicWriters.html
@@ -35,7 +35,7 @@
             <ul>
                 <li><a href="/home"> Home </a></li>
                 <li><a href="/email_validator"> Email Validator </a></li>
-                <li><a href="/jtracker"> Periodic Writers </a></li>
+                <li><a href="/topic_tracker"> Periodic Writers </a></li>
                 <li><a href="/haro_table"> Haro Table </a></li>
                 <li><a href="/contact_us"> Contact Us </a></li>
