+++ flask_app/HTML/topic_tracker.html
@@ -11,7 +11,7 @@
     <link rel="stylesheet" type="text/css" href="{{ url_for('static',filename='css/periodicWriters.css') }}">
     <script src="{{ url_for('static',filename='js/periodicWriters.js') }}" defer></script>
     <link rel="shortcut icon" href="{{ url_for('static',filename='img/favicon.ico') }}">
-    <title>Subscribe to Journalist Updates</title>
+    <title>Subscribe to Topic Updates</title>
 </head>
 
 
@@ -34,7 +34,7 @@
             <ul>
                 <li><a href="/home"> Home </a></li>
                 <li><a href="/email_validator"> Email Validator </a></li>
-                <li><a href="/writers"> Periodic Writers </a></li>
+                <li><a href="/topic_tracker"> Topic Tracker </a></li>
                 <li><a href="/haro_table"> Haro Table </a></li>
                 <li><a href="/contact_us"> Contact Us </a></li>
