+++ topicTrackerRunner.py
@@ -1,5 +1,5 @@
 from flask_app.scripts.create_flask_app import db, app
-from flask_app.scripts.PeriodicWriters.categoricalWriters import send_pdf_report
+from flask_app.scripts.PeriodicWriters.topic_tracker import send_pdf_report
 from flask_app.scripts.PeriodicWriters.weekly import Report
 import pandas as pd
 import traceback
